requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
folium==0.20.0
google-auth==2.40.3
google-genai==1.32.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
Jinja2==3.1.6
joblib==1.5.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.3.2
ortools==9.14.6206
pandas==2.3.2
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
sentry-sdk==2.35.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
tenacity==9.1.2
threadpoolctl==3.6.0
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
xyzservices==2025.4.0
